+++ src/supabase/image-service.ts
@@ -1,9 +1,10 @@
 import { supabase } from "~/supabase/supabase";
 
 export async function uploadImage(file: File) {
+  const uniqueName = `${Date.now()}-${file.name}`;
   const { data, error } = await supabase.storage
     .from("post-images")
-    .upload(`public/${file.name}`, file);
+    .upload(`public/${uniqueName}`, file);
   if (error) {
     console.error("Upload failed", error);
     return null;
